Add support for semerr embedding
Adds the following functions to generated errors:

* As(interface{}) bool
* SetErr(error)

As() is implemented by matching the GRPCCode() and calling SetErr() to
assign the underlying error. No information is lost in the transfer.

The additions allow for embedded semerrs to lend meaning to the parent struct.

The downside is that the parent struct gets a somewhat awkward interface: it
leaks the semerr and has many nested levels to be initialized, likely leading to
a constructor func. It's also worth noting that generated errors now have mixed
receivers.